PM: Deny MPU idle while saving secure RAM
Deny MPU idle before save secure ram and allow it after save secure RAM.
We want to deny MPU going to low power state because, there is a short
time window where a wakeup event would happen around the time the MPU
is going to idle. Since the first thing ROM code does after WFI is to
read INTCPS register, we could reach a situation where the INTCPS is
in idle, but the read is sent to it. That would produce a data abord
during the save of secure ram, which will hang the system. we need
to prevent clock transitions as well during this timeframe.

[nm@ti.com: rebased to 2.6.37-rc2, used omap2_clkdm_deny_idle for
clock prevention]

Patch taken from L-O tree Posted by Nishanth Menon

Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Eduardo Valentin <eduardo.valentin@nokia.com>
Signed-off-by: Deepak K <deepak.k@ti.com>

Change-Id: I05f6623f22b5437587c1485b891a7e5270d33939
Signed-off-by: Steve Schefter <sschefter@ti.com>
Signed-off-by: jack.miller <jack.miller@ti.com>