USB PM: Fix power off mode when USB cable not attached
The board was not coming to power off mode when USB
cable was detached because the OTG_SYSCONFIG register
setting was toggled between smartstandby and forcestandby
in musb_context_save/restore. This new code does not work
on OMAP3630 revisions earlier than ES1.2 since it will
randomly crash when coming out of power off, so older boards
will still not power off when USB is not attached.

Change-Id: I268552ae3244642710064ef397e06512a0345f28
Signed-off-by: Carl Benson <cbenson@windriver.com>
Signed-off-by: Steve Schefter <sschefter@windriver.com>