OMAP3: PM: disable rta
Errata id: i608
RTA (Retention Till Access) feature is not supported and leads to device
stability issues when enabled. This impacts modules with embedded memories
on OMAP3630

Workaround is to disable RTA on boot and coming out of core off.
For disabling rta coming out of off mode, we do this by overriding the
restore pointer for 3630 to allow us restore handler as the first point of
entry before caches are touched and is common for GP and HS devices.
to disable earlier than this could be possible by modifying the ppa for HS
devices.

Ported from 06179768796791c65637607a93acf5fa3c1cbf0b in 2.6.32 by
Girish S G <girishsg@ti.com>

Change-Id: I25f0e094c4a53e35d0d40515ec7fe25ee05fedb7
Signed-off-by: Steve Schefter <sschefter@ti.com>